pylbo: dataseries input checked in ModeVisualisationData
